+++ packages/cloud-toolkit/.eslintrc.js
@@ -1,8 +1,5 @@
 const path = require('path');
 
 module.exports = {
-  extends: path.join(__dirname, '../../.eslintrc.js'),
-  rules: {
-    'no-console': 'off',
-  },
+  extends: path.join(__dirname, '../../.eslintrc.js')
 };