+++ packages/cloud-toolkit/src/types.ts
@@ -6,6 +6,11 @@ import type {
   TSchema,
 } from '@forestadmin/datasource-customizer';
 
+import BootstrapPathManager from './services/bootstrap-path-manager';
+import DistPathManager from './services/dist-path-manager';
+import EventSubscriber from './services/event-subscriber';
+import HttpServer from './services/http-server';
+
 /**
  * This agent the central object used to customize your cloud project
  * some methods available in self-hosted, such as:
@@ -68,3 +73,34 @@ export type EnvironmentVariables = {
   FOREST_SUBSCRIPTION_URL: string;
   FOREST_AUTH_TOKEN: string;
 };
+
+export type MakeCommands = {
+  buildEventSubscriber: BuildEventSubscriber;
+  buildHttpServer: BuildHttpServer;
+  getEnvironmentVariables: () => Promise<EnvironmentVariables>;
+  bootstrapPathManager: BootstrapPathManager;
+  distPathManager: DistPathManager;
+  logger: Logger;
+  login: Login;
+};
+
+export type Spinner = {
+  start: (text?: string) => void;
+  succeed: (text?: string) => void;
+  warn: (text?: string) => void;
+  info: (text?: string) => void;
+  fail: (text?: string) => void;
+  stop: () => void;
+};
+
+export type Logger = {
+  spinner: Spinner;
+  log: (text?: string) => void;
+  error: (text?: string) => void;
+};
+
+export type Login = (logger: Logger) => Promise<void>;
+
+export type BuildHttpServer = (envs: EnvironmentVariables) => HttpServer;
+
+export type BuildEventSubscriber = (vars: EnvironmentVariables) => EventSubscriber;