+++ packages/cloud-toolkit/test/build-commands.test.ts
@@ -1,0 +1,7 @@
+import buildCommands from '../src/build-command';
+
+describe('build commands', () => {
+  it('should be able to execute a command', async () => {
+    expect(() => buildCommands().parse(['help'], { from: 'user' })).not.toThrow();
+  });
+});