+++ packages/cloud-toolkit/test/services/typings-updater.test.ts
@@ -1,5 +1,0 @@
-describe('updateTypings', () => {
-  it('first test', () => {
-    expect(true).toEqual(true);
-  });
-});